LLM/Hahaton/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
openai==1.3.7
python-multipart
asyncssh
requests~=2.32.5